+++ src/lib/vengine/vulkan/renderers/VulkanRendererOutput.cpp
@@ -97,8 +97,8 @@ VkResult VulkanRendererOutput::createGraphicsPipeline()
     VkPipelineVertexInputStateCreateInfo vertexInputInfo = vkinit::pipelineVertexInputStateCreateInfo(0, nullptr, 0, nullptr);
     VkPipelineInputAssemblyStateCreateInfo inputAssembly = vkinit::pipelineInputAssemblyCreateInfo();
 
-    VkViewport viewport = vkinit::viewport(m_swapchainExtent.width, m_swapchainExtent.height, 0.0F, 1.0F);
-    VkRect2D scissor = vkinit::rect2D(m_swapchainExtent.width, m_swapchainExtent.height, 0, 0);
+    VkViewport viewport = vkinit::viewport(static_cast<float>(m_swapchainExtent.width), static_cast<float>(m_swapchainExtent.height), 0.0F, 1.0F);
+    VkRect2D scissor = vkinit::rect2D(     static_cast<int32_t>(m_swapchainExtent.width), static_cast<int32_t>(m_swapchainExtent.height), 0, 0);
     VkPipelineViewportStateCreateInfo viewportState = vkinit::pipelineViewportStateCreateInfo(1, &viewport, 1, &scissor);
 
     VkPipelineRasterizationStateCreateInfo rasterizer =
@@ -175,7 +175,7 @@ VkResult VulkanRendererOutput::createDescriptors(uint32_t imageCount)
 
 VkResult VulkanRendererOutput::updateDescriptors(const VulkanFrameBufferAttachment &inputAttachment)
 {
-    for (size_t i = 0; i < m_descriptorSets.size(); i++) {
+    for (uint32_t i = 0; i < m_descriptorSets.size(); i++) {
         VkDescriptorImageInfo colorAttachmentDescriptor =
             vkinit::descriptorImageInfo(m_inputSampler, inputAttachment.image(i).view(), VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL);
         VkWriteDescriptorSet colorWrite = vkinit::writeDescriptorSet(
