+++ src/lib/vengine/vulkan/renderers/VulkanRendererPost.cpp
@@ -100,8 +100,10 @@ VkResult VulkanRendererPost::createGraphicsPipeline()
     VkPipelineVertexInputStateCreateInfo vertexInputInfo = vkinit::pipelineVertexInputStateCreateInfo(0, nullptr, 0, nullptr);
     VkPipelineInputAssemblyStateCreateInfo inputAssembly = vkinit::pipelineInputAssemblyCreateInfo();
 
-    VkViewport viewport = vkinit::viewport(m_swapchainExtent.width, m_swapchainExtent.height, 0.0F, 1.0F);
-    VkRect2D scissor = vkinit::rect2D(m_swapchainExtent.width, m_swapchainExtent.height, 0, 0);
+    VkViewport viewport =
+        vkinit::viewport(static_cast<float>(m_swapchainExtent.width), static_cast<float>(m_swapchainExtent.height), 0.0F, 1.0F);
+    VkRect2D scissor =
+        vkinit::rect2D(static_cast<int32_t>(m_swapchainExtent.width), static_cast<int32_t>(m_swapchainExtent.height), 0, 0);
     VkPipelineViewportStateCreateInfo viewportState = vkinit::pipelineViewportStateCreateInfo(1, &viewport, 1, &scissor);
 
     VkPipelineRasterizationStateCreateInfo rasterizer =
@@ -174,7 +176,7 @@ VkResult VulkanRendererPost::createDescriptors(uint32_t imageCount,
     VkDescriptorSetAllocateInfo setAllocInfo = vkinit::descriptorSetAllocateInfo(m_descriptorPool, imageCount, setLayouts.data());
     VULKAN_CHECK_CRITICAL(vkAllocateDescriptorSets(m_device, &setAllocInfo, m_descriptorSets.data()));
 
-    for (size_t i = 0; i < imageCount; i++) {
+    for (uint32_t i = 0; i < imageCount; i++) {
         VkDescriptorImageInfo colorAttachmentDescriptor =
             vkinit::descriptorImageInfo(m_inputSampler, colorAttachment.image(i).view(), VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL);
         VkWriteDescriptorSet colorWrite = vkinit::writeDescriptorSet(
